Split JfrFeature into hosted and non-hosted part.
